+++ composer.json
@@ -22,9 +22,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "": "tests"
-        },
-        "classmap": ["tests"]
+            "Opentelemetry\\Tests\\": "test/"
+        }
     },
     "require-dev": {
         "phpunit/phpunit": "^7.5.16",
