+++ src/Exporter/ExporterInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry;
+namespace OpenTelemetry\Exporter;
 
 use OpenTelemetry\Tracing\Span;
 use OpenTelemetry\Tracing\Tracer;
@@ -12,7 +12,7 @@ use OpenTelemetry\Tracing\Tracer;
  *
  * @package OpenTelemetry
  */
-interface Exporter
+interface ExporterInterface
 {
     /**
      * Possible return values as outlined in the OpenTelemetry spec
