+++ src/TransportInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace OpenTelemetry;
 
-interface Transport
+interface TransportInterface
 {
     public function write(array $data) : bool;
 }
